backend/requirements.txt
# ============================================================================
# Backend Dependencies - Sistema de Gestión de Proyectos SVA
# ============================================================================

# FastAPI y servidor ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos y ORM
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validación y configuración
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Telegram Bot
python-telegram-bot==20.7

# Workers asíncronos
celery==5.3.4
celery-redbeat==2.2.0

# Utilidades
python-dotenv==1.0.0
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Linting y formateo (desarrollo)
black==23.12.0
flake8==6.1.0
isort==5.13.1
mypy==1.7.1
